permeate uint for Ids throughout app
not sure I'll stick with this long-term yet, but works for now.

ref https://github.com/PomeloFoundation/Pomelo.EntityFrameworkCore.MySql/issues/601#issuecomment-396346202

+++ src/FilterLists.Data/EntityTypeConfigurations/FilterListTypeConfiguration.cs
@@ -31,7 +31,7 @@ namespace FilterLists.Data.EntityTypeConfigurations
             entityTypeBuilder.Property(x => x.IssuesUrl)
                              .HasColumnType("TEXT");
             entityTypeBuilder.Property(x => x.LicenseId)
-                             .HasDefaultValue(5)
+                             .HasDefaultValue((uint)5)
                              .IsRequired();
             entityTypeBuilder.Property(x => x.Name)
                              .HasColumnType("VARCHAR(126)")
